Break up Turf.jsx to make the code more maintainable
